requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.46
pymysql==1.1.2
requests==2.32.5
jwt==1.4.0
python-dotenv==1.2.1